requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
scikit-learn>=1.4
lightgbm>=4.3
xgboost>=2.0
pyarrow>=15.0
